+++ client/web-sveltekit/src/lib/search/CodeHostIcon.svelte
@@ -13,13 +13,14 @@
     import Tooltip from '$lib/Tooltip.svelte'
 
     export let repository: string
+    export let disableTooltip: boolean = false
 
     $: hostName = repository.split('/')[0]
     $: svgPath = iconMap[hostName]
 </script>
 
 {#if svgPath}
-    <Tooltip tooltip={hostName}>
+    <Tooltip tooltip={disableTooltip ? '' : hostName}>
         <Icon aria-label={hostName} {svgPath} inline />&nbsp;
     </Tooltip>
 {/if}
